Firmware update

+++ src/model/vlan.rs
@@ -1,7 +1,7 @@
 use serde::Deserialize;
 use tabled::Tabled;
 
-use crate::deserialize::null_to_default;
+use crate::util::deserialize::null_to_default;
 
 #[derive(Default, Deserialize)]
 #[serde(default)]
@@ -15,15 +15,15 @@ pub struct Vlan {
     #[serde(deserialize_with = "null_to_default")]
     pub description: String,
     #[serde(deserialize_with = "null_to_default")]
-    pub id: i64,
+    pub id: u64,
     #[serde(deserialize_with = "null_to_default")]
-    pub mod_time: i64,
+    pub mod_time: u64,
     #[serde(deserialize_with = "null_to_default")]
     pub name: String,
     #[serde(deserialize_with = "null_to_default")]
-    pub network_id: i64,
+    pub network_id: u64,
     #[serde(deserialize_with = "null_to_default")]
-    pub node_id: i64,
+    pub node_id: u64,
     #[serde(deserialize_with = "null_to_default")]
-    pub vlanid: i64,
+    pub vlanid: u16,
 }
